Moved tempTitle, GuiEditor, and DevMessage to separate files
